requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Numerical and Scientific Computing
numpy>=1.24.0
scipy>=1.10.0

# Data Visualization
matplotlib>=3.7.0

# Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0

# Training Utilities
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
tensorboard>=2.13.0

# Transformer and Attention Mechanisms
einops>=0.6.1

# Diffusion Models and Generative AI
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0

# Additional ML Tools
wandb>=0.15.0
omegaconf>=2.3.0
